+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.0] - 2022-01-12
 ### Added
 - PauseAudioSource
 - PauseUnityEvent
@@ -22,5 +24,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add initial files
 - Initial commit
 
-[Unreleased]: https://github.com/HyagoOliveira/PauseSystem/compare/0.1.0...main
+[Unreleased]: https://github.com/HyagoOliveira/PauseSystem/compare/1.0.0...main
+[1.0.0]: https://github.com/HyagoOliveira/PauseSystem/tree/1.0.0/
 [0.1.0]: https://github.com/HyagoOliveira/PauseSystem/tree/0.1.0/